Fixed automapper configuration which cause bug where labels could be removed but not added.

+++ PomodoroLibrary/PomodoroLibrary/Data/GeneralRepository.cs
@@ -56,6 +56,9 @@ public class GeneralRepository<T> : IRepository<T> where T : class
 
     public void RemoveRange(IEnumerable<T> models)
     {
-        dbSet.RemoveRange(models);
+        if (models.Count() > 0)
+        {
+            dbSet.RemoveRange(models);
+        }
     }
 }
